Move UserConfig code into seperate files
