+++ src/config.h
@@ -80,6 +80,7 @@ struct Filesystem_config {
     bool extendable;                // is this filesystem extendable? (or read only?)
     bool restorable;                // can a workspace be restored into this filesystem?
 };
+#include "UserConfig.h"
 
 
 // global config, global settings + workspaces
@@ -143,24 +144,6 @@ private:
     void readYAML(string YAML);
 };
 
-
-// config in user home
-class UserConfig {
-private:
-    std::string mailaddress;        // default mailaddress for reminder
-    std::string groupname;          // FIXME: TODO: does this make sense?
-    int reminder;                   // days before expiration to send reminder email
-    int duration;                   // default diration
-
-public:
-    // read config from string, either YAML or single line
-    UserConfig(std::string userconf);
-    // get mailaddress
-    std::string getMailaddress() const { return mailaddress; };
-    std::string getGroupname() const { return groupname; };
-    int getReminder() const { return reminder; };
-    int getDuration() const { return duration; };
-};
 
 
 // helper for std::find
